[ETP]: Fix comparison of narrow type with wide type in loop condition
This should clear up the CodeQL finding in PR #20

+++ isobus/src/can_extended_transport_protocol.cpp
@@ -682,7 +682,7 @@ namespace isobus
 						if (proceedToSendDataPackets)
 						{
 							// Try and send packets
-							for (std::uint8_t i = session->lastPacketNumber; i < session->packetCount; i++)
+							for (std::uint32_t i = session->lastPacketNumber; i < session->packetCount; i++)
 							{
 								dataBuffer[0] = (session->lastPacketNumber + 1);
